Move metrics initialization logic out of constructor
